backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-pymongo==2.3.0
pymongo==4.6.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
gunicorn==21.2.0


